pyproject.toml
[tool.poetry]
name = "sakura"
version = "0.0.1"
description = "Sakura"
authors = []
license = ""
readme = "README.md"
packages = [
    { include = "sakura", from = "src" }
]

[tool.poetry.dependencies]
python = "~3.11"
joblib = "1.4.2"
rich = "14.0.0"
tqdm = { extras = ["all"], version = "^4.66.1" }
requests = "^2.32.0"
ibm-generative-ai = "3.0.0"
javalang = "0.13.0"
toml = "^0.10.2"
jinja2 = "^3.1.2"
openai = "^1.12.0"
python-dotenv = "^1.0.1"
cldk = "1.0.6"
plotly = "^5.22.0"
matplotlib = "^3.9.0"
openpyxl = "3.0.9"
replicate = "^1.0.3"
tree-sitter-java = "^0.23.5"
deepdiff = "^8.0.1"
typer = ">=0.15.3,<0.17.0"
hamster = { git = "git@github.com:aster-test-generation/hamster.git", branch = "feat/include-test-utility-as-helper" }
reaster = { git = "git@github.com:aster-test-generation/re-aster.git", branch = "main" }
javabuild = { git = "git@github.com:aster-test-generation/javabuild.git", branch = "feat/maven-module-scoped-runs" }
langchain = "^0.3.26"
faiss-cpu = "^1.11.0"
langchain-community = "^0.3.26"
langgraph = "^0.5.4"
langchain-openai = "^0.3.30"
google-cloud-aiplatform = "^1.74.0"
pyright = "^1.1.407"
tenacity = "^9.0.0"
python-minifier = "3.2.0"

[tool.poetry.scripts]
sakura = "sakura.cli:app"

[tool.poetry.group.dev.dependencies]
black = "^24.3.0"
pytest = "^8.2.0"
pytest-cov = "^4.1.0"
pylint = "^3.0.2"
flake8 = "^6.1.0"
coverage = "^7.3.2"
pyinstaller = "^6.6.0"
ollama = "^0.5.1"
pyright = "^1.1.407"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "-vv -s --tb=long"